remove captions ui for now
